feat: Add support for resolving TypeScript types in "X-TypeScript-Types" header
It's an opt-in functionality that's only enabled for "?module" queries when "types" query parameter is also set.
The types header contains an unpkg.com URL of the typings file that also has "?module" query parameter set.

In order to fully support environments like "Deno" that expect imported files to be ESM modules, "?modules" transformation is also enabled for the .d.ts, .ts and .tsx files